Refacto removed long in favor of string 12

+++ Sheaft.Domain/ValueObjects/ProductUnitPrice.cs
@@ -6,7 +6,7 @@ public record ProductUnitPrice : UnitPrice
     {
     }
 
-    public ProductUnitPrice(int value) : base(value)
+    public ProductUnitPrice(decimal value) : base(value)
     {
         if(value <= 0)
             throw new InvalidOperationException("Product price must be greater than 0");
